second comit

+++ Milestone-7-Explore-Browser & Debug/Frist-Practice-Day/nosto-bank-main-solution/nosto-bank-main/js/utilities.js
@@ -10,7 +10,7 @@ function getInputFieldValueById(inputFieldId) {
 function getTextElementValueById(elementId) {
   const textElement = document.getElementById(elementId);
   const textElementValueString = textElement.innerText;
-  const textElementValue = parseInt(textElementValueString);
+  const textElementValue = parseFloat(textElementValueString);
   return textElementValue;
 }
